Move fixtures from test_remote.py to conftest
