+++ tests/cli/test_remote.py
@@ -23,79 +23,7 @@ from milatools.cli.remote import (
 )
 from milatools.cli.utils import T, shjoin
 
-from .common import (
-    function_call_string,
-    requires_s_flag,
-    requires_ssh_to_localhost,
-)
-
-
-@pytest.fixture(
-    scope="session",
-    params=[
-        pytest.param("localhost", marks=requires_ssh_to_localhost),
-        # TODO: Think about a smart way to enable this. Some tests won't work as-is.
-        # pytest.param(
-        #     "mila",
-        #     marks=pytest.mark.skipif(
-        #         "-vvv" not in sys.argv, reason="Not testing using the Mila cluster."
-        #     ),
-        # ),
-    ],
-)
-def host(request: pytest.FixtureRequest) -> str:
-    return request.param
-
-
-@pytest.fixture(scope="session")
-def connection(host: str) -> Generator[Connection, None, None]:
-    """Fixture that gives a Connection object that is reused by all tests."""
-    with Connection(host) as connection:
-        yield connection
-
-
-@pytest.fixture(scope="function")
-def MockConnection(
-    monkeypatch: pytest.MonkeyPatch, connection: Connection, host: str
-) -> Mock:
-    """Returns a Mock wrapping the `fabric.connection.Connection` class."""
-    # The return value of the constructor will always be the shared `Connection` object.
-    MockConnection = Mock(
-        name="MockConnection",
-        wraps=Connection,
-        return_value=Mock(
-            name="mock_connection",
-            # Modify the repr so they show up nicely in the regression files and with
-            # consistent/reproducible names.
-            wraps=connection,
-            host=host,
-            __repr__=lambda _: f"Connection({repr(host)})",
-        ),
-    )
-    # mock_connection.configure_mock(
-    #     __repr__=lambda _: f"Connection({repr(host)})",
-    # )
-    import milatools.cli.remote
-
-    monkeypatch.setattr(milatools.cli.remote, Connection.__name__, MockConnection)
-    return MockConnection
-
-
-@pytest.fixture(scope="function")
-def mock_connection(
-    MockConnection: Mock,
-) -> Mock:
-    """returns a Mock wrapping a real `Connection` instance.
-
-    This Mock is used to check how the connection is used by `Remote` and `SlurmRemote`.
-    """
-    mock_connection: Mock = MockConnection.return_value
-    # mock_connection.configure_mock(
-    #     # Modify the repr so they show up nicely in the regression files and with
-    #     # consistent/reproducible names.
-    #     __repr__=lambda _: f"Connection({repr(host)})",
-    # )
-    return mock_connection
+from .common import function_call_string, requires_s_flag
 
 
 @pytest.mark.parametrize("keepalive", [0, 123])
@@ -238,13 +166,6 @@ The command that eventually would be run on the cluster is:
 and `result.stdout.strip()={repr(result.stdout.strip())}`.
 """
     file_regression.check(regression_file_text, extension=".md")
-
-
-@pytest.fixture(scope="function")
-def remote(mock_connection: Connection):
-    hostname = mock_connection.host
-    assert isinstance(hostname, str)
-    return Remote(hostname=hostname, connection=mock_connection)
 
 
 @pytest.mark.parametrize("message", ["foobar"])
